% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{compute.f}
\alias{compute.f}
\title{compute F-score, precision and recall}
\usage{
compute.f(G, H)
}
\arguments{
\item{G}{first label}

\item{H}{second label}
}
\value{
\item{f}{F-score}
\item{p}{precision}
\item{r}{recall}
}
\description{
This fuction is used to compute F-score, precision and recall.
}
\examples{
\dontrun{
res<-compute.f(c1,c2)
}
}
\keyword{F-score,}
\keyword{precision,}
\keyword{recall}
