% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{contigency}
\alias{contigency}
\title{construct contigency matrix}
\usage{
contigency(mem1, mem2)
}
\arguments{
\item{mem1}{first label}

\item{mem2}{second label}
}
\value{
\item{cont}{contigency matrix}
}
\description{
This fuction is used to construct contigency matrix.
}
\examples{
\dontrun{
C<-contigency(mem1,mem2)
}
}
\keyword{contigency}
\keyword{matrix}
