% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deng.r
\docType{data}
\name{deng}
\alias{deng}
\title{deng dataset for demo
including:
deng.label (true cluster labels)
deng.data (deng dataset)}
\format{An object of class \code{list} of length 2.}
\usage{
data(deng)
}
\description{
deng dataset for demo
including:
deng.label (true cluster labels)
deng.data (deng dataset)
}
\examples{
data(deng)
deng.label<-deng$deng.label
deng.data<-deng$deng.data
}
\keyword{dataset}
\keyword{demo}
