% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{evalt}
\alias{evalt}
\title{integrate evalution index}
\usage{
evalt(c1, c2)
}
\arguments{
\item{c1}{first label}

\item{c2}{second label}
}
\value{
\item{res}{list of evaluation results}
}
\description{
This fuction is used to integrate evaluation index.
}
\examples{
\dontrun{
res<-evalt(c1,c2)
}
}
\keyword{evaluation}
