% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{neg}
\alias{neg}
\title{compute negative part of matrix}
\usage{
neg(A)
}
\arguments{
\item{A}{input matrix}
}
\value{
\item{Am}{negative part of input matrix}
}
\description{
This fuction is used to compute negative part of matrix.
}
\examples{
\dontrun{
Am<-neg(A)
}
}
\keyword{negative}
