% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{NMFimpute}
\alias{NMFimpute}
\title{CRNMF for imputation}
\usage{
NMFimpute(data, maxitr = 30, lambda = 1.5, r)
}
\arguments{
\item{data}{data matrix}

\item{maxitr}{max iteration number}

\item{lambda}{tuning parameter}

\item{r}{rank for NMF}
}
\value{
\item{W}{output W of CRNMF}
\item{H}{output H of CRNMF}
\item{S}{output imputation matrix S of CRNMF}
}
\description{
This fuction is used to run CRNMF for imputation.
}
\examples{
\dontrun{
mat<-NMFimpute(data,30,1.5,4)
}
}
\keyword{CRNMF,}
\keyword{imputation}
