% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{nmi}
\alias{nmi}
\title{compute normalized mutual information}
\usage{
nmi(label, result)
}
\arguments{
\item{label}{first label}

\item{result}{second label}
}
\value{
\item{v}{normalized mutual information}
}
\description{
This fuction is used to compute normalized mutual information.
}
\examples{
\dontrun{
res<-nmi(c1,c2)
}
}
\keyword{information}
\keyword{mutual}
\keyword{normalized}
