% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{NNDSVD}
\alias{NNDSVD}
\title{run NNDSVD for initialization}
\usage{
NNDSVD(A, k, flag)
}
\arguments{
\item{A}{input matrix}

\item{k}{rank}

\item{flag}{mode for NNDSVD}
}
\value{
\item{W}{initial W for following NMF}
\item{H}{initial H for following NMF}
}
\description{
This fuction is used to run NNDSVD for initialization.
}
\examples{
\dontrun{
ini<-NNDSVD(data,30,0)
}
}
\keyword{NNDSVD}
