% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{pos}
\alias{pos}
\title{compute positive part of matrix}
\usage{
pos(A)
}
\arguments{
\item{A}{input matrix}
}
\value{
\item{Ap}{positive part of input matrix}
}
\description{
This fuction is used to compute positive part of matrix.
}
\examples{
\dontrun{
Ap<-pos(A)
}
}
\keyword{positive}
