% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{rand.idx}
\alias{rand.idx}
\title{compute Rand index}
\usage{
rand.idx(c1, c2)
}
\arguments{
\item{c1}{first label}

\item{c2}{second label}
}
\value{
\item{AR}{adjusted Rand index}
\item{RI}{Rand index}
}
\description{
This fuction is used to compute Rand index.
}
\examples{
\dontrun{
res<-rand.idx(c1,c2)
}
}
\keyword{Rand}
\keyword{index}
