% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{select.r}
\alias{select.r}
\title{select rank}
\usage{
select.r(data)
}
\arguments{
\item{data}{data matrix}
}
\value{
\item{r}{selected rank}
}
\description{
This fuction is used to select rank.
}
\examples{
\dontrun{
k<-select.r(data)
}
}
\keyword{rank}
