% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNMF.R
\name{soft.threshold}
\alias{soft.threshold}
\title{compute soft threshold}
\usage{
soft.threshold(A, gamma)
}
\arguments{
\item{A}{data matrix}

\item{gamma}{tuning parameters}
}
\value{
\item{S}{matrix output}
}
\description{
This fuction is used to compute soft threshold.
}
\examples{
\dontrun{
A<-soft.threshold(data,1.5)
}
}
\keyword{soft}
\keyword{threshold}
